+++ CHANGELOG.rst
@@ -14,7 +14,12 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
-[2.5.5] - 2021-01-05
+[2.5.6] - 2021-01-06
+* Updated the StudentProctoredExamAttempt view's PUT handler to allow for a 
+  new action "mark_ready_to_resume", which transitions exam attempts in the "error" state
+  to a "ready_to_resume" state.
+
+[2.5.5] - 2020-01-05
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Cover `Start System Check` button on the proctoring instruction page with the
   conditions software download link is provided by the proctoring provider,